Moved scanning setup dialog to the root of the pos from home as it used in 2 places now
