+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/scanningsetup/WooPosScannerDetectionServiceForTracking.kt
@@ -1,8 +1,8 @@
-package com.woocommerce.android.ui.woopos.home.scanningsetup
+package com.woocommerce.android.ui.woopos.scanningsetup
 
 import com.woocommerce.android.ui.woopos.common.util.ScannerInfo
 import com.woocommerce.android.ui.woopos.common.util.WooPosScannerDetectionUtil
-import com.woocommerce.android.ui.woopos.home.scanningsetup.WooPosScanningSetupState.BarcodeReaderDevice
+import com.woocommerce.android.ui.woopos.scanningsetup.WooPosScanningSetupState.BarcodeReaderDevice
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
